langchainprj/pyproject.toml
[project]
name = "langchainprj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "IPython",
    "accelerate>=0.24.1",
    "bitsandbytes>=0.41.3",
    "chromadb>=1.2.1",
    "gradio>=6.13.0",
    "huggingface-hub",
    "ipykernel",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.1.0",
    "langchain-ollama>=0.0.4",
    "langchain-openai>=1.2.1",
    "numpy>=1.24.0",
    "plotly>=5.14.3",
    "pypdf>=6.10.2",
    "python-dotenv>=1.0.0",
    "ruff>=0.15.12",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=3.0.1",
    "torch>=2.0.0",
    "transformers>=4.43.0",
]

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cuda" }

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
